refactor(profile): moved components into modules
