+++ lib/components/cover_widget.dart
